refactor(rule): use tabs to select test target and split components
