+++ src/views/RuleEngine/components/SQLTest.vue
@@ -8,133 +8,82 @@
     </div>
     <el-collapse-transition>
       <div v-show="isTesting">
-        <div class="vertical-align-center radio-group-container">
-          <label> {{ tl('testTarget') }} </label>
-          <el-radio-group v-model="testTarget" @change="handleTestMethodChanged">
-            <el-radio-button :label="TestRuleTarget.SQL">SQL</el-radio-button>
-            <el-radio-button :label="TestRuleTarget.Rule">{{ tl('rule') }}</el-radio-button>
-          </el-radio-group>
-        </div>
-        <div class="vertical-align-center radio-group-container" v-if="isTestRule">
-          <label>{{ tl('inputData') }}</label>
-          <el-radio-group v-model="inputData" @change="handleTestMethodChanged">
-            <el-radio-button :label="InputData.Mock">{{ tl('mockData') }}</el-radio-button>
-            <el-radio-button :label="InputData.Real">{{ tl('realData') }}</el-radio-button>
-          </el-radio-group>
-        </div>
-        <div class="test-header" v-if="isMockInput">
-          <label class="test-label">
-            {{ tl('dataSource') }}
-            <InfoTooltip :content="tl('dataSourceDesc')" />
-          </label>
-          <p v-if="isDataTypeNoMatchSQL" class="no-match-tip">
-            {{ tl('dataTypeSQLNoMatch') }}
-          </p>
-          <FromSelect
-            v-model="dataType"
-            :ingress-bridge-list="ingressBridgeList"
-            :event-list="eventList"
-            for-test
-          />
-        </div>
-        <el-divider />
-        <el-row class="input-output-row" :gutter="26">
-          <!-- Input data -->
-          <el-col :span="12" v-if="isMockInput">
-            <label class="test-label">
-              {{ tl('testData') }}
-              <InfoTooltip :content="tl('testDataDesc')" />
-            </label>
-            <el-card shadow="never" class="test-card with-border">
-              <TestSQLContextForm v-model="testParams.context" />
-            </el-card>
-          </el-col>
-          <el-col :span="12" v-else>
-            <label class="test-label">{{ tl('testingWithRealData') }}</label>
-            <el-card shadow="never" class="test-card with-border tip-card">
-              <p class="tip" v-if="!isTestStarted">{{ tl('pleaseClickStartTest') }}</p>
-              <p class="tip" v-else>{{ tl('waitingRealData') }}</p>
-            </el-card>
-          </el-col>
-          <!-- Output -->
-          <el-col :span="12" v-if="!isTestRule">
-            <label class="test-label" shadow="none">
-              {{ tl('outputResult') }}
-              <InfoTooltip :content="tl('outputResultDesc')" />
-            </label>
-            <el-card class="test-result">
-              <div class="result-hd">
-                <el-tooltip effect="dark" placement="top" :content="t('Base.copy')">
-                  <el-button type="primary" class="btn-copy" link>
-                    <el-icon class="icon-copy" :size="18" @click="copyText(resultData)">
-                      <copy-document />
-                    </el-icon>
-                  </el-button>
-                </el-tooltip>
-              </div>
-              <div class="monaco-container">
-                <Monaco
-                  :id="createRandomString()"
-                  v-model="resultData"
-                  class="payload"
-                  lang="json"
+        <label> {{ tl('testTarget') }} </label>
+        <!-- @tab-change="handleTestMethodChanged" -->
+        <el-tabs v-model="testTarget">
+          <el-tab-pane label="SQL" :name="TestRuleTarget.SQL" lazy>
+            <div>
+              <div class="test-header">
+                <label class="test-label">
+                  {{ tl('dataSource') }}
+                  <InfoTooltip :content="tl('dataSourceDesc')" />
+                </label>
+                <p v-if="isDataTypeNoMatchSQL" class="no-match-tip">
+                  {{ tl('dataTypeSQLNoMatch') }}
+                </p>
+                <FromSelect
+                  v-model="dataType"
+                  :ingress-bridge-list="ingressBridgeList"
+                  :event-list="eventList"
+                  for-test
                 />
               </div>
-            </el-card>
-          </el-col>
-          <el-col :span="12" v-else>
-            <label class="test-label" shadow="none">
-              {{ tl('outputResult') }}
-              <InfoTooltip :content="tl('outputResultDesc')" />
-            </label>
-            <el-card class="test-result rule-result">
-              <el-scrollbar :max-height="490" ref="ScrollbarCom">
-                <div class="collapse-wrap">
-                  <LogDataDisplay :log-data="logData" />
-                </div>
-              </el-scrollbar>
-            </el-card>
-          </el-col>
-        </el-row>
-        <div class="buttons-bar">
-          <el-button
-            v-if="!isTestRule && isMockInput"
-            type="primary"
-            :loading="testLoading"
-            plain
-            :icon="CaretRight"
-            @click="submitTestSQL"
-          >
-            {{ tl('testsql') }}
-          </el-button>
-          <template v-if="isTestRule">
-            <template v-if="!isTestStarted">
-              <div class="btn-start-container">
+              <el-row class="input-output-row" :gutter="26">
+                <el-col :span="12">
+                  <label class="test-label">
+                    {{ tl('testData') }}
+                    <InfoTooltip :content="tl('testDataDesc')" />
+                  </label>
+                  <el-card shadow="never" class="test-card with-border">
+                    <TestSQLContextForm v-model="testParams.context" />
+                  </el-card>
+                </el-col>
+                <el-col :span="12" v-if="!isTestRule">
+                  <label class="test-label" shadow="none">
+                    {{ tl('outputResult') }}
+                    <InfoTooltip :content="tl('outputResultDesc')" />
+                  </label>
+                  <el-card class="test-result">
+                    <div class="result-hd">
+                      <el-tooltip effect="dark" placement="top" :content="t('Base.copy')">
+                        <el-button type="primary" class="btn-copy" link>
+                          <el-icon class="icon-copy" :size="18" @click="copyText(resultData)">
+                            <copy-document />
+                          </el-icon>
+                        </el-button>
+                      </el-tooltip>
+                    </div>
+                    <div class="monaco-container">
+                      <Monaco
+                        :id="createRandomString()"
+                        v-model="resultData"
+                        class="payload"
+                        lang="json"
+                      />
+                    </div>
+                  </el-card>
+                </el-col>
+              </el-row>
+              <div class="buttons-bar">
                 <el-button
                   type="primary"
+                  :loading="testLoading"
                   plain
-                  @click="startTest"
                   :icon="CaretRight"
-                  :disabled="!savedAfterRuleChange"
+                  @click="submitTestSQL"
                 >
-                  {{ tl('startTest') }}
+                  {{ tl('testsql') }}
                 </el-button>
-                <p class="tip" v-if="!savedAfterRuleChange">{{ tl('pleaseSaveFirst') }}</p>
+                <el-button plain :icon="RefreshRight" @click="resetContext">
+                  {{ t('Base.reset') }}
+                </el-button>
               </div>
-            </template>
-            <template v-else>
-              <el-button v-if="isMockInput" type="primary" plain @click="submitTestRule">
-                {{ tl('submitTest') }}
-              </el-button>
-              <el-button plain @click="stopTest">
-                {{ tl('stopTest') }}
-              </el-button>
-            </template>
-          </template>
-          <el-button v-if="isMockInput" plain :icon="RefreshRight" @click="resetContext">
-            {{ t('Base.reset') }}
-          </el-button>
-        </div>
+            </div>
+          </el-tab-pane>
+          <el-tab-pane :label="tl('rule')" :name="TestRuleTarget.Rule" lazy>
+            <RuleTest :key="createRandomString()" :rule-data="ruleData" />
+          </el-tab-pane>
+        </el-tabs>
       </div>
     </el-collapse-transition>
   </div>
@@ -142,28 +91,22 @@
 
 <script setup lang="ts">
 import { testsql } from '@/api/ruleengine'
-import { createRandomString, getKeywordsFromSQL, waitAMoment } from '@/common/tools'
+import { createRandomString, getKeywordsFromSQL } from '@/common/tools'
 import InfoTooltip from '@/components/InfoTooltip.vue'
 import Monaco from '@/components/Monaco.vue'
-import useDebugRule, { useStatusController } from '@/hooks/Rule/rule/useDebugRule'
+import { useStatusController } from '@/hooks/Rule/rule/useDebugRule'
 import { useRuleUtils } from '@/hooks/Rule/rule/useRule'
 import useCopy from '@/hooks/useCopy'
-import useDataNotSaveConfirm from '@/hooks/useDataNotSaveConfirm'
 import useI18nTl from '@/hooks/useI18nTl'
 import { RuleInputType, TestRuleTarget } from '@/types/enum'
 import { BridgeItem, RuleEvent } from '@/types/rule'
 import { CaretRight, CopyDocument, RefreshRight } from '@element-plus/icons-vue'
 import { ElMessage, ElMessageBox } from 'element-plus'
 import JSONbig from 'json-bigint'
-import { PropType, Ref, computed, defineProps, onUnmounted, ref, watch } from 'vue'
+import { PropType, Ref, computed, defineProps, onMounted, ref, watch } from 'vue'
 import FromSelect from '../components/FromSelect.vue'
-import LogDataDisplay from './LogDataDisplay.vue'
 import TestSQLContextForm from './TestSQLContextForm.vue'
-
-const enum InputData {
-  Mock = 'mock',
-  Real = 'real',
-}
+import RuleTest from './RuleTest.vue'
 
 const JSONbigNative = JSONbig({ useNativeBigInt: true })
 
@@ -196,7 +139,7 @@ const props = defineProps({
 const ruleSql = computed(() => props.ruleData?.sql || '')
 
 const isTestRule = computed(() => testTarget.value === TestRuleTarget.Rule)
-const { isTesting, savedAfterRuleChange, testTarget } = useStatusController()
+const { isTesting, testTarget } = useStatusController()
 
 const testLoading = ref(false)
 const resultData = ref<string>('')
@@ -351,84 +294,11 @@ const setDataTypeNContext = () => {
   testParams.value = { context, output: '' }
 }
 
-const inputData = ref<InputData>(InputData.Mock)
-const isMockInput = computed(
-  () => testTarget.value === TestRuleTarget.SQL || inputData.value === InputData.Mock,
-)
-
-const {
-  logData,
-  emptyLogArr,
-  handleStopTest,
-  startTestRuleUseMockData,
-  submitMockDataForTestRule,
-  startTestRuleUseRealData,
-  setCbAfterPolling,
-} = useDebugRule()
-
-const ScrollbarCom = ref()
-const scrollLogToBottom = async (log: string) => {
-  if (log) {
-    let isScrollToBottom = false
-    const scrollWrap = ScrollbarCom.value.wrapRef
-    const scrollContent = scrollWrap?.firstChild
-    if (scrollWrap && scrollContent) {
-      isScrollToBottom =
-        Math.abs(scrollWrap.clientHeight + scrollWrap.scrollTop - scrollContent.clientHeight) < 5
-    }
-    await waitAMoment()
-    if (isScrollToBottom) {
-      ScrollbarCom.value.scrollTo({
-        top: scrollContent.clientHeight - scrollWrap.clientHeight,
-        behavior: 'smooth',
-      })
-    }
-  }
-}
-
-const isTestStarted = ref(false)
-const startTest = async () => {
-  isTestStarted.value = true
-  try {
-    if (isMockInput.value) {
-      await startTestRuleUseMockData(props.ruleData.id)
-      // TODO:TODO:TODO:TODO:confirm
-      await waitAMoment(1000)
-      await submitMockDataForTestRule(props.ruleData.id, getMockContext())
-    } else {
-      await startTestRuleUseRealData(props.ruleData.id)
-    }
-    setCbAfterPolling(scrollLogToBottom)
-  } catch (error) {
-    //
-  }
-}
-const submitTestRule = async () => {
-  try {
-    await submitMockDataForTestRule(props.ruleData.id, getMockContext())
-  } catch (error) {
-    //
-  }
-}
-const handleTestMethodChanged = () => {
-  emptyLogArr()
-  stopTest()
-}
-const stopTest = () => {
-  handleStopTest()
-  isTestStarted.value = false
-}
-
-const judgeNeedRemindUser = () => isTestStarted.value
-useDataNotSaveConfirm(judgeNeedRemindUser, 'RuleEngine.debugLeaveConfirm')
-
-setDataTypeNContext()
-
-watch(() => props.ruleData, stopTest)
-
-onUnmounted(() => {
+onMounted(() => {
   isTesting.value = false
 })
+
+setDataTypeNContext()
 </script>
 
 <style lang="scss">
